+++ packages/gov-libs/utils/getCENNZnetApi.ts
@@ -1,7 +1,7 @@
 import type { CENNZNetNetwork } from "@cennznet/api/types";
 
 import { Api } from "@cennznet/api";
-import { CENNZ_NETWORK } from "@gov-libs/lib/constants";
+import { CENNZ_NETWORK } from "@gov-libs/constants";
 
 export async function getCENNZnetApi(): Promise<Api> {
 	return await Api.create({
